# docs/swagger/docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Cart management API for e-commerce platform",
    "title": "Cart API",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "description": "Returns a list of products with pagination",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "List products",
        "operationId": "listProducts",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "Page number for pagination",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "Number of items per page",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Filter by SKU",
            "name": "sku",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Filter by name",
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of products",
            "schema": {
              "$ref": "#/definitions/ProductListResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "Creates a new product in the system",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Products"
        ],
        "summary": "Create a product",
        "operationId": "createProduct",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateProductParams"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product created",
            "schema": {
              "$ref": "#/definitions/ProductResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateProductParams": {
      "type": "object",
      "required": [
        "sku",
        "name",
        "price",
        "inventory"
      ],
      "properties": {
        "inventory": {
          "description": "Available inventory",
          "type": "integer"
        },
        "name": {
          "description": "Product name",
          "type": "string"
        },
        "price": {
          "description": "Product price",
          "type": "number",
          "format": "float"
        },
        "sku": {
          "description": "Product SKU",
          "type": "string"
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "required": [
        "message",
        "code",
        "server_time"
      ],
      "properties": {
        "code": {
          "description": "Error code",
          "type": "string"
        },
        "data": {
          "description": "Additional error data",
          "type": "object",
          "x-nullable": true
        },
        "message": {
          "description": "Error message",
          "type": "string"
        },
        "server_time": {
          "description": "Server timestamp",
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "ProductListResponse": {
      "allOf": [
        {
          "$ref": "#/definitions/StandardResponse"
        },
        {
          "type": "object",
          "properties": {
            "data": {
              "type": "object",
              "properties": {
                "products": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "description": "Product ID",
                        "type": "string",
                        "format": "uuid"
                      },
                      "inventory": {
                        "description": "Available inventory",
                        "type": "integer"
                      },
                      "name": {
                        "description": "Product name",
                        "type": "string"
                      },
                      "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "float"
                      },
                      "sku": {
                        "description": "Product SKU",
                        "type": "string"
                      }
                    }
                  }
                },
                "total": {
                  "description": "Total number of products",
                  "type": "integer"
                }
              }
            }
          }
        }
      ]
    },
    "ProductResponse": {
      "allOf": [
        {
          "$ref": "#/definitions/StandardResponse"
        },
        {
          "type": "object",
          "properties": {
            "data": {
              "type": "object",
              "properties": {
                "id": {
                  "description": "Product ID",
                  "type": "string",
                  "format": "uuid"
                },
                "inventory": {
                  "description": "Available inventory",
                  "type": "integer"
                },
                "name": {
                  "description": "Product name",
                  "type": "string"
                },
                "price": {
                  "description": "Product price",
                  "type": "number",
                  "format": "float"
                },
                "sku": {
                  "description": "Product SKU",
                  "type": "string"
                }
              }
            }
          }
        }
      ]
    },
...
    "StandardResponse": {
      "type": "object",
      "required": [
        "data",
        "message",
        "code",
        "server_time"
      ],
      "properties": {
        "code": {
          "description": "Response code",
          "type": "string"
        },
        "data": {
          "description": "Response data payload",
          "type": "object"
        },
        "message": {
          "description": "Response message",
          "type": "string"
        },
        "server_time": {
          "description": "Server timestamp",
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}